fix: give the 404 page its own title; drop unused Google Fonts preconnects
